# specs/001-java-backend/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: Arca Java Backend API
  version: "1.0.0"
  description: |
    REST API for ArcaDigitalis Vault. Provides SIWE authentication, non-custodial
    transaction payloads, Lit ACC generation/validation, event indexing, artifact
    storage, and notification subscription management.

    **Security**: All endpoints except `/auth/**`, `/config`, `/packages/{packageKey}/status`,
    `/packages/{packageKey}/recovery-kit`, `/packages/{packageKey}/tx/renew`,
    `/packages/{packageKey}/tx/rescue`, `/acc-template`, and `/health/**` require
    a valid Bearer JWT issued by `POST /auth/verify`.

    **Non-custodial**: All `**/prepare` and `**/payload` endpoints return unsigned
    calldata. The backend never signs or submits transactions.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    StoredArtifact:
      type: object
      required: [id, artifactType, sha256Hash, sizeBytes, createdAt]
      properties:
        id:                  { type: string, format: uuid }
        artifactType:        { type: string, enum: [manifest, ciphertext] }
        sha256Hash:          { type: string }
        ipfsUri:             { type: string, nullable: true }
        s3Uri:               { type: string, nullable: true }
        sizeBytes:           { type: integer, format: int64 }
        createdAt:           { type: string, format: date-time }
        storageConfirmedAt:  { type: string, format: date-time, nullable: true }
    NotificationTarget:
      type: object
      required: [id, chainId, proxyAddress, packageKey, subscriberAddress, eventTypes, channelType, channelValue, active]
      properties:
        id:                  { type: string, format: uuid }
        chainId:             { type: integer, format: int64 }
        proxyAddress:        { type: string }
        packageKey:          { type: string }
        subscriberAddress:   { type: string }
        eventTypes:          { type: array, items: { type: string } }
        channelType:         { type: string, enum: [email, webhook] }
        channelValue:        { type: string }
        active:              { type: boolean }
        createdAt:           { type: string, format: date-time }
        lastDeliveryAttempt: { type: string, format: date-time, nullable: true }
        lastDeliveryStatus:  { type: string, enum: [delivered, failed, retrying], nullable: true }
paths:
...
  /artifacts:
    post:
      summary: Pin manifest or ciphertext artifact
      operationId: pinArtifact
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [artifactType, sha256Hash, content]
              properties:
                artifactType: { type: string, enum: [manifest, ciphertext] }
                sha256Hash:   { type: string, description: "Expected sha256 (0x-prefixed hex)" }
                chainId:      { type: integer, format: int64 }
                proxyAddress: { type: string }
                packageKey:   { type: string }
                content:      { type: string, format: binary }
      responses:
        "201":
          description: Artifact pinned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoredArtifact' }
        "422": { description: sha256 mismatch, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /artifacts/{id}:
    get:
      summary: Retrieve artifact metadata by ID
      operationId: getArtifact
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoredArtifact' }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ─── Event Index ─────────────────────────────────────────────────────────────
...
  /notifications/subscriptions:
    post:
      summary: Create notification subscription
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chainId, proxyAddress, packageKey, eventTypes, channelType, channelValue]
              properties:
                chainId:       { type: integer, format: int64 }
                proxyAddress:  { type: string }
                packageKey:    { type: string }
                eventTypes:    { type: array, items: { type: string }, minItems: 1 }
                channelType:   { type: string, enum: [email, webhook] }
                channelValue:  { type: string }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationTarget' }
        "403":
          description: Caller is not the on-chain owner or beneficiary of the package
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /notifications/subscriptions/{id}:
    put:
      summary: Update notification subscription
      operationId: updateSubscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventTypes:   { type: array, items: { type: string } }
                channelValue: { type: string }
                active:       { type: boolean }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationTarget' }
        "403": { description: Not the subscription owner, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

    delete:
      summary: Delete notification subscription
      operationId: deleteSubscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "403": { description: Not the subscription owner, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ─── Health ──────────────────────────────────────────────────────────────────
...